refactor: refactor name of config type file
