+++ src/components/StickerCreator/stickerBackgroundColors.ts
@@ -1,4 +1,4 @@
-import type { ProductCardConfig } from '../../types/product.ts';
+import type { ProductCardConfig } from '../../types/preparedProductData.ts';
 import { theme } from '../../theme.ts';
 
 export const stickerBackgroundColor: Record<ProductCardConfig['color'], string> = {
