+++ src/features/catalog/components/productCard/ProductCard.tsx
@@ -6,7 +6,7 @@ import IconButton from '@mui/material/IconButton';
 import Box from '@mui/material/Box';
 import { CardActions } from '@mui/material';
 import AddShoppingCartRoundedIcon from '@mui/icons-material/AddShoppingCartRounded';
-import type { ProductCardConfig } from '../../../../types/product';
+import type { ProductCardConfig } from '../../../../types/preparedProductData.ts';
 import ShelterSticker from '../../../../components/StickerCreator/ShelterSticker';
 import { formatPrice } from '../../../../utils/formatPrice/formatPrice';
 import styles from './ProductCard.module.scss';
